2 Forks into One...
Instead of working on both this fork and the CanopyMC AreaTP project,
I've started merging the best bits of both into this fork.

I have already re-written this fork from the ground up, and now this
fork doesn't even resemble the original upstream. The original
ClaimTeleport project is a Java learning experience for the dev, and
it's better for him if I don't spoon feed him everything, otherwise he
won't be learning anything.